third_party/WebKit/LayoutTests/fast/css/invalidation/invalidation-set-with-adjacent-combinators-expected.txt

Check that targeted class invalidation works with when adjacent combinators are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(i1, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(i1, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(i1, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

This text should have a green background.
